Changed cross-platform backend resolution for Linux
On linux, the backend is now selected based off of the
`XDG_CURRENT_DESKTOP` environment variable. Before this, I defaulted
for Gtk (for some reason). If both are installed on the machine, then
it would choose Gtk when it should have choosen Qt. This is a better
solution. The list of desktop environments is not complete, but it
is a good start.