clipboard: improved image handling
Added code in gtkbackend to make image handling more explicit. Now,
`set_image` doesn't take a format. Instead, it is found out implicitly
via the `isinstace` function. Also, an image converter can be passed
to the method so that if the image is invalid, it can be converted and
passed again.

The same converter principle was added to the `set_image` method as
well. Now, if a format is invalid and a converter is given, then it
will return that type of image.