events: created event-based asynchronous framework
Created a new event-driven framework that all (or most) clipboard
backends should implement. There is an abstract event listener that
each backend has. Each has an event handler and a clipboard. The
start method is an awaitable coroutine that is abstract. The event
listener polls for events from the clipboard backend and uses the
handler to handle them. The event loop and dispatcher should be
whatever the native clipboard uses.
Events are subclassed from an abstract event class. The user defines
his or her own events.
The clipboard class has an instance of the listener task. It passes
itself and an event handler.